Change home page
Change some links and feature descriptions on the home page.
Also generate the sitemap and add a more high level description on home page for people who might not be familiar with Companies House data products. And fixed the bug in the footer showing undefined. Added an open source license. Added a link to disqualifications.

Signed-off-by: Brian Evans <ebrian101@gmail.com>